fix(Dockerfile): use npm for starting the application
When running the application with npm, the environment variable
NODE_ENV is set to production and npm_package_version is set to the
version in package.json. This is not the case when running simply with
node. This commit fixes this issue and now in the docker image and
in docker compose we use `npm run start:prod`.